moved exceptions into new exception package
