+++ backend/src/main/java/p3/group/p3_aau_football/exceptions/DocumentNotFoundException.java
@@ -1,4 +1,4 @@
-package p3.group.p3_aau_football.statistic.common;
+package p3.group.p3_aau_football.exceptions;
 
 public class DocumentNotFoundException extends RuntimeException {
     public DocumentNotFoundException(String message) {
