+++ backend/src/main/java/p3/group/p3_aau_football/statistic/league/LeagueStatisticsNotFoundException.java
@@ -1,6 +1,6 @@
 package p3.group.p3_aau_football.statistic.league;
 
-import p3.group.p3_aau_football.statistic.common.DocumentNotFoundException;
+import p3.group.p3_aau_football.exceptions.DocumentNotFoundException;
 
 public class LeagueStatisticsNotFoundException extends DocumentNotFoundException {
     public LeagueStatisticsNotFoundException(String message) {
